docs/tutorials/websocket-source/requirements.txt
python-dateutil
websockets
quixstreams~=3.3.0
